Fix segfault in capture card editor
There can be a segmentation fault in the capture card editor
of mythtv-setup when configuring DVB-S satellite cards.
This happens when a capture card is changed from DVB-S,with
a configured input type of LNB (anything but not Unconnected),
to another type of DVB card.
After the capture card has been changed to a non-DVB-S type and
the DiSEqC tree has been deleted, a "Save then Exit" refers
to the already deleted DiSEqC tree. This creates a segmentation fault.
There are many more things to fix in the DiSEqC configuration
code and once this is done the "delete m_root" can be enabled again.
For now it is better to have a memory leak than a segmentation fault.

Refs #13465