No more channel conflicts on ATSC major/minor numbers.
This change is similar to that for channel conflicts on service ID's for DVB.

Rationale (also the same):
A "conflicting channel" in mythtv-setup is a channel that does not have a unique channel number.
It is essential to have a unique channel number for all channels in a video source otherwise
mythbackend can arbitrarily select any channel with the same channel number for a recording.
However, mythtv-setup also considers a channel conflicting if the ATSC major/minor number
combination is the same as another channel and then forces the user to enter a unique channel number.

In my understanding this is a bug because A. it does not change the ATSC major/minor numbers
and B. ATSC major/minor number combinations do not need to be unique in MythTV.
For ATSC the test on the major/minor number combination uniqueness is removed and all ATSC channels
are now considered "non-conflicting" to start with.
The ATSC major/minor number combination is used to create default channel numbers
but later on the check on the uniqueness of the channel number is still done.

Refs #13472