Full search for old channels
In mythtv-setup page Channel Scan there is now a new
option "Full search for old channels".

If set, compare all channels in the database with the channels found in
the scan; otherwise only the channels in the same transport are compared.
This option is useful when you want to keep channel data such as
the xmltvid and the icon path when doing a rescan
after channels have been rearranged across transports.

For ATSC the comparison is done only on the major and minor numbers.
For DVB the comparison of channels in the same transport is done
on service id but the comparison of channels in different transports
is done on network id and service id.

Without this option, when channels have moved to another transport
the old channels are classified as "off-air" and can only be deleted.
The moved channels are classified as "New" and they then do not have
the non-scanned data anymore such as the xmltvid and the icon path
that may have been added to the old channel.

The option is currently unchecked by default. If this option proves to
be a succes then this can be changed to checked by default.

Refs #13472