MythTV: Add compatibility to python3 for '__repr__' methods
The 'repr()' method calls the '__repr__' method of an instanciated class in
python and is used mostly in the interactive python shell, like

$ python2
Python 2.7.15+ (default, Nov 27 2018, 23:36:35)

>>> from MythTV import MythDB, Recorded
>>> db = MythDB()
>>> rec = Recorded((3002, 20190622125000), db = db)
>>> rec     # -----> Note: this calls __repr__ under the hood.
b'<Recorded 'Reisezelt - Kurztrip','2019-06-47 12:50:00+02:00' at 0x7fcdb0faabe0>'
>>>

With Python3, this 'repr()' method returns a 'str' type (i.e. unicode),
in contrast to Python2, where this method may return a native 'str' type (ascii)
or an 'utf-8' encoded string.

Using this in a Python3 shell, leads to the following Type-Error:

  $ python3
  Python 3.6.8 (default, Jan 14 2019, 11:02:34)
  >>> from MythTV import MythDB, Recorded
  >>> db = MythDB()
  >>> rec = Recorded((3002, 20190622125000), db = db)
  >>> rec
  Traceback (most recent call last):
    File "<stdin>", line 1, in <module>
  TypeError: __repr__ returned non-string (type NoneType)
  >>>

This patch adds compatibility to Python3, it simply selects the current
python version when returning data from the '__repr__()' method.