Number of transports in mythtv-setup log output
The number of transports in the summary of the log output
is now always the number of transports found in the scan.
In the processing of the scan results, at some point all
transports of channels in the database that are not found in the scan,
called the off-air channels, are added to the list of transports.
This caused an incorrect number of transports to be presented.
This is now fixed.
Also added a summary of the scan parameters in the log output.
Also added, when option -v chanscan is given, a separate
listing of all transports found, in order of ascending frequency.

Refs #13472