Update recording starttime for single recordings
When a program guide update from the EIT changes the
starttime of a program and if that program is sceduled
to be recorded as a single reording then update
the starttime of hte scheduled recording.

If the starttime of a scheduled single recording is
not correct then the program is shown as "Not Listed"
in the Scheduled Recording page because the starttime
is used to identify the program to be recorded.

Fixes #13481