EITCache check on table version
The EITcache code that checks if an EIT event contains new data or
that it contains data that has been already processed did now allow
for missing table versions when the table version wrapped around
from 31 to 0.
If that happens then the guide data of that channel will not
be updated for a few days until the table version is 31 again.
Note that it is completely valid for a receiver to not receive
all table versions. This happens for instance when a tuner is
used for recording; during that time the active EIT scanning
is suspended and only the EIT data on the multiplex of the
channel that is currently being reorded is processed.
This happens also when a device is switched off.
The check as now implemented only checks if the new table version
is different from the previously processed table version.

(cherry picked from commit 1ad17d49ea6bdfb16c3cd5e4882227a32f06ef2a)
Signed-off-by: Klaas de Waal <kdewaal@mythtv.org>