mythfrontend: Convert deinterlacer settings to new format
- NOTE - the database has not been changed. The existing deinterlacer
preferences in the video display profile have been re-purposed and now
represent the quality and OpenGL/driver preferences.
- for example, an old preference might have been 'vdpauadvanced' but
would now read high:driver or high:shader:driver
- this helps with switching between the development branch and master
for the time being

- new settings seem to be working correctly -  though the presentation
and/or wording may be changed.

- known issues with the new code:
    - there is currently no fallback to other deinterlacers when CPU
deinterlacers are unavailable but GLSL and/or driver are not selected.
This happens with most 'copy-back' hardware decoders that return NV12
frames - which are not supported by libavfilter deinterlacing.
    - timestretch is not dropping back to single rate deinterlacing when
necessary - which usually manifests as broken a/v sync and no audio