Silence warning on exit of python if a mysql cursor is in use
On python3, if you call one of the 'databaseSearch' methods
like in the following sequence 'searchRecorded', you get a
traceback on exit() like this one:

    $ python
    Python 3.6.8 (default, Aug 20 2019, 17:12:48)
    >>> from MythTV import MythDB, Recorded, Program
    >>> db = MythDB()
    >>> reciter = db.searchRecorded(basename = u'3002_201906....mkv')
    >>> rec = next(reciter)
    >>> rec
    <Recorded 'Reisezeit - Kurzflip','2019-06-22 12:50:00+02:00' at..>
    >>> prgm = rec.getRecordedProgram()
    >>> prgm
    <RecordedProgram 'Reisezeit - Kurzflip','2019-06-22 12:54:07.....>
    >>> prgm.title
    'Reisezeit - Kurzflip'
    >>> prgm.videoprop
    'WIDESCREEN,AVC,720'
    >>> exit()
    Exception ignored in: <generator object databaseSearch.__call__ at..>
    Traceback (most recent call last):
      File "..../MythTV/utility/other.py", line 185, in __call__
      File "..../MythTV/_conn_mysqldb.py", line 110, in __exit__
      File "..../MythTV/_conn_mysqldb.py", line 102, in rollback
    _mysql_exceptions.OperationalError:
                                     (2006, 'MySQL server has gone away')

Note: For 'databaseSearch' methods, see
https://www.mythtv.org/wiki/0.25_Python_Bindings/Connection_Handlers
for searchRecorded, searchOldRecorded, searchVideos etc.

Some background info:
MythTV's python bindings use a set of connections to the mysql database,
managed by the class 'DBConnection' and it's parent class
'_Connection_Pool'.
The latter one has a method to delete all aquired connections during
garbage collection ('__del__'), which unconditionally closes all open
connections to the database (opened via the python package MySQLdb,
derived from https://github.com/PyMySQL/mysqlclient-python).

On python2, this '__del__' method is never called on exit, on python3,
this method gets called because of improvements on the garbage collector
(See PEP 442).
Unfortunately, this garbage collection kicks in too early and the opened
cursor claimed by one of the 'databaseSearch' methods gets closed without
finishing the last action ('commit' or 'rollback').
See the tracaback mentioned above.

I verified (see '(*)') that these open connections are closed on exit
by the garbage collection of the package MySQLdb itself, therefore there
is no need to close these connections within the class 'DBConnection' of
the MythTv's bindings explicitely.

(*) Verified by code inspection (in MySQLdb) and observing the
mysql command 'show processlist' as root user in another terminal.

The implemented solution is simple and works for python2 and python3:
Rename the '__del__' method in order to not beeing called by the
pythons garbage colletor of python3.

Final note:
This commit inhibits only a 'warning' on shutdown of a python terminal,
it does not change any behaviour of the python bindings, because the
'databaseSearch' methods only reads from the 'mythconverg' database,
therefore 'commit' or 'roolback' does not do anything on that database.