AvFormatDecoder: Use a single FFmpeg thread for hardware decoding
- also use asynchronous callbacks for software decoding. All of our
get_formant and get_buffer calls shouls be thread safe and per the FFmpeg
documentation, this should improve concurrency.

- the single thread for hardware decoding is needed as using
asynchronous callbacks means we can get callbacks later than needed.
When using AVHWFramesContext (VDPAU and VAAPI) with H264 in particular,
the decoder resets the hardware context when seeking - but the callback
to release our context is not received until after we have created the
new context; which leads to large spikes in memory consumption and
potential playback errors.