Fix TestIPTVRecorder after change to IPTVChannelInfo
In a60e7429a09a ("IPTVChannelInfo: explictly set IPTVTuningData protocol to
invalid") I removed `IPTVChannelInfo::IsValid()` thinking it was unused but
missed the uses in these test cases. Reintroduce the method, but make it
`protected` and a friend of the `TestIPTVRecorder` class.

The above change also changed the behaviour of `IPTVChannelInfo`'s constructors
such that the protocol is not always probed, and therefore `IsValid()` is not
necessarily true. This is the case when constructed via
`IPTVChannelFetcher::ParsePlayList`. Update `TestIPTVRecorder::ParseChanInfo` to
reflect this by now asserting that the tuning data is not valid and that the
protocol is `inValid` as expected.

Refs #13488

Signed-off-by: Paul Harrison <pharrison@mythtv.org>