VAAPI/VDPAU: Workaround pause frame retention when seeking
- both VAAPI and VDPAU use AVHWFramesContexts when direct rendering. For
H264 content, the hardware context is recreated when we flush the decoder
but the old context is not released (due to reference counting) until we
have released the retained hardware pause frame.
- this causes temporary spikes in video resource usage and can lead to
failures - particularly with VDPAU.
- so ensure the pause frame is released when using VDPAU/VAAPI and H264
- this means there is no pause frame when seeking - so there will often
be an unwelcome screen blanking.
- the only viable alternative is to force OpenGLVideo to use the resize
stage for these decoder/codec combinations and use the intermediate
framebuffer as a 'pause frame of last resort'.
- there are other pause frame related issues when seeking but not yet
sure of cause.