Add DVB constants
Add private descriptors for Freesat strams on Astra 28.2E satellite.
Add private data specifier ID for Freesat. The presence of this value
in a private data specifiers indicates that one or more of
the following descriptors are Freesat private descriptors.
Add table with original network ID's. This allows use of
symbolic names instead of numbers in code that checks network ID's.