Freesat logical channel numbers
Get the channel number of the Freesat streams on the
Astra 28.2E satellite from the BAT for the selected
bouquet ID and the selected region.
The defaults for the bouquet ID and the region ID can be overridden
with the environment variables FSAT_BOUQUET_ID and FSAT_REGION_ID
awaiting a database schema update after which the values
can be selected in mythtv-setup page "Video source".