Fix resource leak in VAAPI and VDPAU
- both VAAPI and VDPAU FFmpeg hardware device contexts set the free and
user_opaque members (not documented) when using the default
av_hwdevice_ctx_create functionality
- we override/reset these to track proper desctruction/reference
counting and to access the OpenGL interopability object.
- as a result, VAAPI and VDPAU hardware contexts were not finishing
properly and release their devices.
- this did not manifest with VAAPI as we use our own initialisation for
the bulk of the time but it caused playback errors with VDPAU after a
number of seeks - as there were multiple concurrent (and leaked) VDPAU
devices.
- we now store the default free/user_opaque values in MythOpenGLInterop
and ensure they are called as the last possible operation before
hardware device destruction
- this also requires a default/dummy MythOpenGLInterop implementation
that does not require an OpenGL context but is used to store those
values when checking functionality and/or not direct rendering (i.e.
decode only).
- also fixes a typo that prevented VDPAU decode only from working.