AvFormatDecoder: Refactor OpenFile for better live tv support
- default, non-livetv behaviour should be largely unchanged
- livetv startup in testing is now faster and more reliable

- codec initialisation is now nested in 2 loops
- the inner loop attempts to open the input stream for up to 5seconds,
resetting the ringbuffer after each attempt to ensure all available data
is used for the scan.
- when complete, the second/outer loop will try to ensure the necessary
stream parameters are available - again forcing a retry when it fails.
The AVFormatContext is closed before trying again to avoid errors
regarding re-opening codecs.
- certain audio parameter checks are disabled as they often fail
(problem here is often the Audio Description Channel) but subsequent
initialisation is fine.