util-nvctrl: Fix NV-CONTROL refresh rate detection
- this appears to have stopped working correctly, in my case because the
incorrect display mask was being returned.
- according to the headers NV_CTRL_CONNECTED_DISPLAYS is deprecated -
which may account for the change in behaviour
- so convert to the recommened usage, the primary benefit being we
retrieve an actual display identifier (e.g. DFP-2) from the libarary and
use this to filter the metamodes.
- I also took the opportunity to welcome the code to the world of Qt C+
+11 (rather than a copy and paste from the nvidia settings code).
- sundry changes include simpler string parsing, cleanup of sundry clang
warnings, move if/def'd code out of DisplayResX, better debugging of
rates/modes, header cleanup and filtering out of unsupported modes that
are also not returned by xrandr.
- with NV-CONTROL in use and -v playback logging, the final set of modes/
rates will be logged in a format similar to the output of xrandr
- code still seems clunky and could probably be simplified further