Fix sundry display issues when using fullscreen and/or resolution switching
- update VideoOutWindow to only process changes when a parameter has
actually changed. This avoids unnecessary signals to the video classes
and attempts to stop spamming the logs with updates.
- focus on the window size in VideoOutWindow as we are always bound by
the window size - not the display size - and other areas of the code
need to ensure the window is resized and VideoOutWindow updated as
necessary
- 'display visible rect' is not pertinent to OpenGLVideo - the window
rect is king for the OpenGL view port (we previously used display
visible rect as a bad proxy as it was never updated).
- reimplement WindowResized in VideoOutputOpenGL to ensure we pick up
window resizing events.
- sundry removal of unused methods/members
- NB there are still various issues. OSD font sizing when the window has
been resized is erratic for some reason (both when switching to
fullscreen and using display resolution switching). When using display
resolution switching, switching to the guide grid etc is not working
quite right and needs more work.