Services: Dvr/GetSavedBookmark - better handling for when there is no seek table
If there is no seektable, GetSavedBookmark with duration option is not able to
calculate the duration. In that case it was returning a garbage value
(from an uninitialized variable). Changed so it will return -1 in that case.

(cherry picked from commit 36cf11e9f6941fcd3bfba7acc369ca008b86db65)