Astra-2 28.2E satellites BSkyB channel numbers
In addition to the Freesat channel numbers now also the BSkyB
channel numbers can be retrieved from the private descriptors
on a channel scan of the Astra-2 28.2E satellites.
Freesat and BSkyB each have a number of bouquets that contain
the descriptors with the channel numbers.
Selection is done with the bouquet ID and additionally
with a Freesat- or BSkyB-specific region ID.
Currently the bouquet ID and the region ID can be set
with environment variables FSAT_BOUQUET_ID and FSAT_REGION_ID.