tidy: Eliminate unnecessary copying of objects.
The clang-tidy "unnecessary value param" check pointed out a number of
places where objects were passed by copying when they could be passed
by const reference.  The vast majority of these fixes convert a
QString object into a const QString& reference.  There are also a
number of other objects that get passed by reference instead of
copying.

These changes resulted in the removal of a number of calls to
'::std::move' that were added six months ago.  They essentially
propagate the use of references to a point where the variables are
still in use and can no longer be moved.

https://clang.llvm.org/extra/clang-tidy/checks/performance-unnecessary-value-param.html
https://clang.llvm.org/extra/clang-tidy/checks/modernize-pass-by-value.html