Selection of Bouquet and Region in Video sources page
Add fields bouquet_id and region_id to table videosource for
database schema version 1351. This will be done when the database
schema version is changed in mythversion.h.
Conditional on the database schema version bigger than 1350
are the following changes:
- Add selection of the Bouquet ID and the Region ID in the Video sources page of mythtv-setup.
- In the channelscan, use the bouquet_id and region_id values from table videosource.
If the database version is 1350 the old code is still used.
After the database version has been changed to 1351 (or higher)
the  old code can be removed.