tidy: Make sure all local variables are initialized. (libmythui/upnp)
The clang-tidy "local variable initial value" checker pointed out a
large number of uninitialized local variables in functions.  Where
possible these were fixed by reducing the scope of the variable and
combining the declaration and initialization of the variable into a
single statement.  When that wasn't possible, suitable initial values
were provided for the variable.  In one or two instances there are
minor code tweaks to not assign an initial value and then immediately
reassign the same value in the 'else' clause of an 'if' statement.

https://clang.llvm.org/extra/clang-tidy/checks/cppcoreguidelines-init-variables.html