tidy: Silence false 'redundant expression' warnings.
The clang-tidy 'redundant expression' checker pointed out several
places in the goom code that appear to have redundant expressions.
These uses of the RAND() inline macro are not redundant because a side
effect of the function is to increment a pointer into an array, so
each call actually returns a different value.

https://clang.llvm.org/extra/clang-tidy/checks/misc-redundant-expression.html