tidy: Silence false 'missing comma' warnings in long database SQL statements.
The clang-tidy 'missing comma' checker looks for places where a list
of strings is used, and attempts to determine if there is a comma
missing after any of the items in the list.  This fails when looking
at very long SQL statements that have been broken onto multiple lines
for easier reading.

https://clang.llvm.org/extra/clang-tidy/checks/bugprone-suspicious-missing-comma.html