Fix issues with default schedgroup and reclimit values.
My database had somehow been changed such that schedgroup was
defaulting to enabled but wasn't for others.  That is fixed now.

Also, change the default reclimit officially back to 1.  With
schedgroup enabled, the reclimit is automatically increased when
needed.  For those few cases, where schedgroup can not be used, the
user can set the desired reclimit when disabling schedgroup.

Fixes #13501