tidy: Silence false 'misleading indentation' warnings in a couple of places.
The clang-tidy 'misleading indentation' checker looks for places where
the code syntax and its indentation don't match up, and flags these
for review.  It seems to have a problem with an 'emit' call after a
if/then/else clause as three of these are incorrectly flagged.  There
are a;sp two places where multiple if/ifdef compilation directives are
used that throw off the indentation (depending upon which defines are
used to compile the program).

https://clang.llvm.org/extra/clang-tidy/checks/readability-misleading-indentation.html