tidy: Fix some "order of initialization" problems.
The clang-tidy "order of initialization" checker looks for global
initializers that reference external objects.  Since the C++ Standard
doesn't define the order in which global initializers are called, its
possible that these initializers may be called before the initializers
for the things they reference.

This commit contains trivial fixes to this order of initialization
problem.  It simply moves code around inside of the same cpp file to
eliminate the use of the external keyword.

https://clang.llvm.org/extra/clang-tidy/checks/cppcoreguidelines-interfaces-global-init.html