Suggested channel number in case of conflicts
When there are after a channel scan duplicate channel numbers
a suggested channel number is presented to the user.
The algorithm to create this number is now slightly changed.
For ATSC the algorithm is not changed.
For DVB the sequence is now:
- Try the existing channel number
- Try the service ID
- If there is a Frequency ID (channel) try frequency ID - service ID
- Try Service ID - Network ID
- Try Service ID - Transport ID
- Find an unused channel number.
This has the following advantages:
- Duplicate channels are now in the Channel Editor sorting adjacent
to the original channel. This makes it easy to select which one to keep.
- The unused channel numbers are kept unused as long as possible;
allocating a channel to an unused channel number can cause another
duplicate later in the scanning process.
Note that this problem happens in practice only when scanning
satellites; usually the service ID, which is used as channel number
when there is no Logical Channel Number available, is unique.

Refs #13472