Default scanfrequency in videosource
Add a default scanfrequency in mythtv-setup page "Video source".
This frequency is used as default for the "Full Scan (Tuned)"
options of the channel scanner for DVB-T/T2, DVB-C and DVB-S/S2.
If there is already a multiplex for the selected video source with
that frequency then the values from the multiplex are used
as defaults for the "Full Scan (Tuned)" parameters.

Refs #13742