tidy: Silence a couple of misleading "infinite loop" warnings.
The clang-tidy "infinite loop" checker pointed out a couple of places
that look like infinite loops, because the loop variable isn't updated
anywhere inside of the loop.  All of these are written in the form:

    while ( x=fn() )
        {...}

What clang-tidy doesn't see is that inside the function is an iterator
that will eventually return 0 and cause the loop to exit.  Tell
clang-tidy to ignore these loops in the future.

https://clang.llvm.org/extra/clang-tidy/checks/bugprone-infinite-loop.html