Show transport ID and frequency in Channel Options editor
In mythtv-setup, show the transport stream ID and the frequency of the
transport stream in the Channel Editor, page "Channel Options - Common".
Add and extend some help texts in this page.
The transport stream ID or the frequency is needed to find
the transport stream of a channel in the "Edit Transports" page.

Refs #13472