tidy: Return braced initializer instead of creating object just for return.
The clang-tidy "return braced initializer list" check pointed out two
places where explicitly creating an object and then immediately
returning it to the caller could be replaced by just returning a
braces initializer for the object.  This eliminates the duplication of
the type information in the function.  Changes made by the clang-tidy
program.

https://clang.llvm.org/extra/clang-tidy/checks/modernize-return-braced-init-list.html