Fix segfault in mythtv-setup Transport Editor
This fixes a regression in the mythtv-setup Transport Editor,
introduced in commit 952e76461c62c22f3383e7639cef39344d94e70e.
The new classes/fields TransportID and Frequency, used in the
Channel Editor / Channel Options - Common page, are given a value
elsewhere in the code.
However, these classes do need a Load() function to override the
default Load() function in StandardSettings. This prevents a
segfault in the Transport Editor, although these new classes
are not used in the Transport Editor.