Hz/kHz confusion in SatelliteDeliverySystemDescriptor
In SatelliteDeliverySystemDescriptor the function FrequencyHz
did return the frequency in kHz. This is, although misleading,
not a real bug because almost everwhere the value was correctly
interpreted as a frequency in kHz, as is standard for all
frequencies in the context of satellites.
Only in the channel scanner, where the NIT table is parsed
for additional transports, the result of FrequencyHz was
divided by 1000 and this resulted in a value in MHz.
Contrary to what the title of ticket #13058 suggests,
this did not prevent finding and scanning new multiplexes.
The code uses the frequency value to search for an existing
multiplex and if it cannot find it then a new multiplex is created.

Fixes #13058

Signed-off-by: Klaas de Waal <kdewaal@mythtv.org>