Fix GUI issues in mythtv-setup Transport Editor
Various cleanups needed after recent changes to the mythtv-setup GUI.
Revert commit 49d4eebd339d99b78b62fe1685b231c2c500d3df which did
address the segfault but did not address the root cause.
The root cause of segfault was the adddtion of a "class Frequency"
to show the frequency in the Channel Options page.
This conflicted with the "class Frequency" that already existed
in the Transport Editor page.
Class names must be unique but there is no diagnostic for this.
Also:
- Disable selection of video source in Transport Editor; this
  must be done in the previous Channel Editor page.
- Improvement of help texts.