tidy: Fix a couple of error checks on calls to posix_fadvise.
The clang-tidy "posix return" checker pointed out a couple of places
where the return value of a function call is incorrectly tested for an
error.  The test is for a negative number, but the function returns
zero for success and a positive number on error.  Change these test
check for a non-zero value instead of a negative value.

https://clang.llvm.org/extra/clang-tidy/checks/bugprone-posix-return.html