tidy: Clean up code stripping dvd or bluray device names.
The clang-tidy bugprone "branch clone" checker pointed out two places
where it appears that the bodies of an if/then/else clause were
duplicated.  It appears that this code was designed to strip a device
name from the front of the string, and replace a double "//" with a
single slash.  Rewrite these two clauses to simplify them.

https://clang.llvm.org/extra/clang-tidy/checks/bugprone-branch-clone.html