tidy: Clean up duplicated conditional if/then/else and switch bodies.
The clang-tidy bugprone "branch clone" checker pointed out two places
where the bodies of an if/then/else clause appeared to be duplicated
because the code inside of the clause was only conditionally compiled.
Making the entire "else if" clause conditionally compiled (instead of
just the body) eliminates most of these warnings.  The remaining
warnings are truly duplicated code, so flag them to be ignored by
clang-tidy.

https://clang.llvm.org/extra/clang-tidy/checks/bugprone-branch-clone.html