tidy: Clean up duplicated if/then/else statement in lcdprocclient.cpp.
The clang-tidy bugprone "branch clone" checker pointed out a spot
where the bodies of an if/then/elseif statement were duplicated.
Given that the first test is for "x==1", and the second test for
"x!=0" is also true when the first test is true, the entire first
clause can be dropped without changing the functionality.

https://clang.llvm.org/extra/clang-tidy/checks/bugprone-branch-clone.html