On-screen display of time stretch in playback
Reduce floating point inaccuracy in the time stretch value
caused by repeatedly changing the value.
This is the patch proposed by amb@gedanken.org.uk but with
a numeric constant for the time stretch step value so that
it is defined in only one place.
Change the presentation of the time stretch value
to show always two decimal places.

Fixes #12973